% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{rmdcountAddin}
\alias{rmdcountAddin}
\title{rmdcountAddin}
\usage{
rmdcountAddin()
}
\value{
nothing
}
\description{
Applies \code{rmdcount} to the current R Markdown document
}
\examples{
if (interactive()) rmdcountAddin()
}
